requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
fastapi
uvicorn
google-cloud-storage
xgboost
